requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
Flask-HTTPAuth==4.8.0
Flask-SQLAlchemy==3.1.1
fpdf==1.7.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.4.0
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.32.3
SQLAlchemy==2.0.31
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
